+++ src/main/java/com/ramanbabich/investment/limetraderclient/marketdata/model/Candle.java
@@ -29,4 +29,4 @@ public record Candle(
     BigDecimal high,
     BigDecimal low,
     BigDecimal close,
-    Integer volume) {}
+    Long volume) {}
